make it easier to configure your own cdn
